refactor: :art: move MXBI utils to separate module and reorganize imports
