+++ src/pymxbi/mxbi/__init__.py
@@ -1,6 +1,6 @@
-from .mxbi import MXBI, get_mxbi, set_mxbi
 from .factory import MXBIModel, build_mxbi
-
+from .mxbi import MXBI
+from .utils import get_mxbi, set_mxbi
 
 __all__ = [
     "MXBI",
